fixed bug reset password
