+++ store/src/main/java/store/configuration/controller/UserController.java
@@ -3,19 +3,13 @@ package store.configuration.controller;
 import java.security.Principal;
 import java.sql.Timestamp;
 import java.util.Date;
-import java.util.Locale;
 import java.util.Optional;
-import java.util.UUID;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.MessageSource;
-import org.springframework.core.env.Environment;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -37,7 +31,6 @@ import store.configuration.model.User;
 import store.configuration.model.UserVerificationToken;
 import store.configuration.registration.OnRegistrationCompleteEvent;
 import store.configuration.service.CommentService;
-import store.configuration.service.PasswordResetTokenService;
 import store.configuration.service.UserVerificationTokenService;
 import store.configuration.service.impl.UserService;
 
@@ -54,14 +47,6 @@ public class UserController {
 	private UserVerificationTokenService tokenService;
 	@Autowired
 	private UserDetailsService manager;
-	@Autowired
-	private MessageSource messages;
-	@Autowired
-	private JavaMailSender mailSender;
-	@Autowired
-	private Environment env;
-	@Autowired
-	private PasswordResetTokenService passwordResetToken;
 	private Date date = new Date();
 
 	@RequestMapping(value = "/logout", method = { RequestMethod.GET, RequestMethod.POST })
@@ -120,7 +105,8 @@ public class UserController {
 	}
 
 	@RequestMapping(path = "/comment/edit/{id}", method = RequestMethod.POST)
-	public String editComment(@Valid @ModelAttribute("comment") Comment comment, BindingResult result, Principal principal) {
+	public String editComment(@Valid @ModelAttribute("comment") Comment comment, BindingResult result,
+			Principal principal) {
 		if (result.hasErrors()) {
 			return "error";
 		}
@@ -128,67 +114,6 @@ public class UserController {
 		comment.setAuthor(user);
 		commentService.saveComment(comment);
 		return "redirect:/object/list/";
-	}
-
-	@RequestMapping(value = "/resetPassword", method = RequestMethod.GET)
-	public ModelAndView displayResetPassword(ModelAndView modelAndView, String email) {
-		modelAndView.addObject("email", email);
-		modelAndView.setViewName("forgotPassword");
-		return modelAndView;
-	}
-
-	@RequestMapping(value = "/resetPassword", method = RequestMethod.POST)
-	public String resetPassword(final HttpServletRequest request, final Model model,
-			@RequestParam("email") String userEmail) {
-		final User userFind = userService.findByEmail(userEmail);
-		if (userFind == null) {
-			model.addAttribute("message", "This email does not exist!");
-			return "error";
-		}
-		final String token = UUID.randomUUID().toString();
-		userService.createPasswordResetTokenForUser(userFind, token);
-		final SimpleMailMessage email = constructResetTokenEmail(getAppUrl(request), request.getLocale(), token,
-				userFind);
-		mailSender.send(email);
-
-		model.addAttribute("message", "Request to reset password received. Check your inbox for the reset link.");
-		return "successForgotPassword";
-	}
-
-	@RequestMapping(value = "/changePassword", method = RequestMethod.GET)
-	public String changePassword(final HttpServletRequest request, final Model model, @RequestParam("id") final long id,
-			@RequestParam("token") final String token) {
-		final String result = passwordResetToken.validatePasswordResetToken(id, token);
-		if (result != null) {
-			model.addAttribute("message", "result null");
-			return "redirect:/user/login";
-		}
-		return "redirect:/:/user/resetPassword";
-
-	}
-
-	@RequestMapping(value = "/savePassword", method = RequestMethod.POST)
-	public String savePassword(final HttpServletRequest request, final Model model,
-			@RequestParam("password") final String password) {
-		final User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		userService.changeUserPassword(user, password);
-		model.addAttribute("message", "Passwrod save");
-		return "redirect:/user/login";
-	}
-
-	private SimpleMailMessage constructResetTokenEmail(String contextPath, Locale locale, String token, User user) {
-		String url = contextPath + "/user/changePassword?token=" + token;
-		String message = messages.getMessage("message.resetPassword", null, locale);
-		return constructEmail("Reset Password", message + " \r\n" + url, user);
-	}
-
-	private SimpleMailMessage constructEmail(String subject, String body, User user) {
-		SimpleMailMessage email = new SimpleMailMessage();
-		email.setSubject(subject);
-		email.setText(body);
-		email.setTo(user.getEmail());
-		email.setFrom(env.getProperty("support.email"));
-		return email;
 	}
 
 }
