+++ store/src/main/resources/messages.properties
@@ -1,0 +1,3 @@
+messages.properties
+messages_en_US.properties
+messages_ru_RU.properties
