+++ store/src/main/webapp/WEB-INF/views/forgotPassword.html
@@ -8,7 +8,7 @@
 </head>
 
 <body>
-	<form action="#" th:action="@{/user/resetPassword/}" method="post">
+	<form action="#" th:action="@{/resetPassword/}" method="post">
 		<table>
 			<tr>
 				<td><label for="email">Email</label></td>
