+++ store/src/main/webapp/WEB-INF/views/login.html
@@ -23,7 +23,7 @@
 		<button class="btn btn-lg btn-primary btn-block" name="Submit"
 			value="Login" type="Submit" th:text="Login"></button>
 	</form>
-	<a href="/store/user/resetPassword">Forgot Password?</a>
+	<a href="/store/resetPassword">Forgot Password?</a>
 	<a href="/store/user/registration">Registration</a>
 	<span th:text="${message}"></span>
