+++ store/src/main/webapp/WEB-INF/views/resetPassword.html
@@ -8,19 +8,13 @@
 </head>
 <body>
 	<h2>Enter new password:</h2>
-	<form action="#" th:action="@{/user/savePassword}" th:object="${user}"
-		method="post">
+	<form action="#" th:action="@{/savePassword}" method="post">
+	
 		<table>
 			<tr>
-				<td><label for="email">Email</label></td>
-				<td><input th:field="*{email}" type="text" name="emailId"
-					readonly></input></td>
-			</tr>
-			<tr>
-				<td><label for="password">Password</label></td>
-				<td><input th:field="*{password}" type="password"
-					name="password"></input></td>
-			</tr>
+			<td><label for="password">password</label></td>
+				<td><input type="password" id="password" name="password"
+					th:placeholder="password" class="form-control" />
 			<tr>
 				<td><input type="reset" value="Clear" /></td>
 				<td><input type="submit" value="Submit"></input></td>
